Address review feedback on the VRT MAX provider
Split helpers.py, which had grown past 1600 lines, into models, queries,
parsers, api_client and auth modules, and move the station table into
constants.py. The GraphQL documents live in their own module rather than
inflating the client with 340 lines of query text.

Errors now use Music Assistant's own types: VrtApiError is a
ResourceTemporarilyUnavailable, VrtNotFoundError a MediaNotFoundError and
VrtAuthError a LoginFailed. Most of the catch-and-re-raise glue in the provider
goes away with it, and a transient failure is no longer reported as "not found".

The supported features are no longer a fixed set. The podcast library features
are declared only when credentials are configured, so an instance without an
account no longer offers sync switches that cannot work - the enabled by
default "sync library deletions" would otherwise prune previously synced
podcasts on the first sync that legitimately found nothing. A failed token
refresh now aborts the favourites sync for the same reason, rather than
reporting an empty one. supported_media_types declares radio so the provider
takes part in radio matching and single media type searches. Without an
account the on-demand folders are hidden and podcasts and episodes are marked
unplayable, instead of being offered and refused once playback is attempted,
and credentials are verified during setup rather than at first play.

Episode positions ran backwards: VRT lists newest first while Music Assistant
numbers oldest to newest, so rank_episodes_by_date now assigns them. Programme
pages, episode lists and the browse listings are cached, the programme page is
read once instead of twice, and the resume-point target moved from a
hand-rolled dictionary on the client to @use_cache, so it survives a restart.

Requests are paced by a Throttler and the tracklist fetches run under a
TaskManager limit, which replaces the invented fifty episode ceiling. Only the
newest ten tracklists are fetched, since the chapters are rendered in the now
playing view. Search runs its two catalogue queries together under a deadline
inside the eight seconds a provider search is given. Only the DRM-free HLS
rendition is accepted; falling back to another target returned a manifest that
could never play. Artwork is requested as a sized rendition rather than the
full-resolution original, which a browse list fetched by the megabyte.

Login uses the shared client session helper, so it gets the connection pool,
DNS resolver and the real user agent, and token expiry is read with PyJWT
rather than hand-decoded from base64 with a thirty minute guess on failure.
Live radio plays the AAC stream where VRT offers one, at the same 128kbps as
the MP3 it replaces. Search and browse each log what they were asked for and
how much they returned.
