+++ tests/providers/vrt_max/test_playback.py
@@ -2,13 +2,26 @@
 
 from __future__ import annotations
 
+import logging
+from typing import Any, Self
 from unittest.mock import AsyncMock, Mock
 
 import pytest
 from music_assistant_models.enums import MediaType, StreamType
-from music_assistant_models.errors import MediaNotFoundError, UnplayableMediaError
+from music_assistant_models.errors import (
+    LoginFailed,
+    MediaNotFoundError,
+    ResourceTemporarilyUnavailable,
+    UnplayableMediaError,
+)
+from music_assistant_models.media_items import (
+    ItemMapping,
+    PodcastEpisode,
+    ProviderMapping,
+)
 
-from music_assistant.providers.vrt_max.helpers import (
+from music_assistant.providers.vrt_max.api_client import VrtMaxClient
+from music_assistant.providers.vrt_max.models import (
     VrtApiError,
     VrtAuthError,
     VrtProgress,
@@ -18,6 +31,30 @@ from music_assistant.providers.vrt_max.helpers import (
 from music_assistant.providers.vrt_max.provider import VrtMaxProvider
 
 EPISODE_ID = "/vrtmax/podcasts/radio-1/h/pod/1/1--ep/"
+
+
+def _episode(duration: int) -> PodcastEpisode:
+    """Build a minimal PodcastEpisode, as MA hands one to on_played."""
+    return PodcastEpisode(
+        item_id=EPISODE_ID,
+        provider="vrt_max--test",
+        name="Ep",
+        position=1,
+        duration=duration,
+        provider_mappings={
+            ProviderMapping(
+                item_id=EPISODE_ID,
+                provider_domain="vrt_max",
+                provider_instance="vrt_max--test",
+            )
+        },
+        podcast=ItemMapping(
+            media_type=MediaType.PODCAST,
+            item_id="/vrtmax/podcasts/radio-1/h/pod/",
+            provider="vrt_max--test",
+            name="Pod",
+        ),
+    )
 
 
 async def test_episode_stream_details_requires_auth(provider: VrtMaxProvider) -> None:
@@ -52,7 +89,7 @@ async def test_episode_stream_details_success(provider: VrtMaxProvider) -> None:
 
 
 async def test_episode_stream_details_auth_error(provider: VrtMaxProvider) -> None:
-    """A player-token auth failure maps to UnplayableMediaError."""
+    """A player-token auth failure surfaces as LoginFailed."""
     provider._auth.enabled = True  # type: ignore[misc]
     provider._client.get_stream_info = AsyncMock(  # type: ignore[method-assign]
         return_value=VrtStreamInfo("streamid", 1800)
@@ -61,18 +98,18 @@ async def test_episode_stream_details_auth_error(provider: VrtMaxProvider) -> No
         side_effect=VrtAuthError("bad creds")
     )
 
-    with pytest.raises(UnplayableMediaError):
+    with pytest.raises(LoginFailed):
         await provider.get_stream_details(EPISODE_ID, MediaType.PODCAST_EPISODE)
 
 
 async def test_episode_stream_details_api_error(provider: VrtMaxProvider) -> None:
-    """A stream-info API failure maps to MediaNotFoundError."""
+    """A transient stream-info failure surfaces as ResourceTemporarilyUnavailable."""
     provider._auth.enabled = True  # type: ignore[misc]
     provider._client.get_stream_info = AsyncMock(  # type: ignore[method-assign]
         side_effect=VrtApiError("boom")
     )
 
-    with pytest.raises(MediaNotFoundError):
+    with pytest.raises(ResourceTemporarilyUnavailable):
         await provider.get_stream_details(EPISODE_ID, MediaType.PODCAST_EPISODE)
 
 
@@ -141,7 +178,7 @@ async def test_on_played_fully_played_writes_total_duration(provider: VrtMaxProv
         return_value=target
     )
     provider._client.post_resume_point = AsyncMock()  # type: ignore[method-assign]
-    media_item = Mock(duration=1800)
+    media_item = _episode(duration=1800)
 
     await provider.on_played(
         MediaType.PODCAST_EPISODE,
@@ -165,7 +202,7 @@ async def test_on_played_partial_writes_actual_position(provider: VrtMaxProvider
         return_value=target
     )
     provider._client.post_resume_point = AsyncMock()  # type: ignore[method-assign]
-    media_item = Mock(duration=1800)
+    media_item = _episode(duration=1800)
 
     await provider.on_played(
         MediaType.PODCAST_EPISODE,
@@ -181,7 +218,7 @@ async def test_on_played_partial_writes_actual_position(provider: VrtMaxProvider
 async def test_on_played_ignores_non_episode_media_type(provider: VrtMaxProvider) -> None:
     """Progress is only written back for podcast episodes."""
     provider._auth.enabled = True  # type: ignore[misc]
-    media_item = Mock(duration=1800)
+    media_item = _episode(duration=1800)
 
     await provider.on_played(
         MediaType.RADIO, "radio1", fully_played=True, position=10, media_item=media_item
@@ -194,7 +231,7 @@ async def test_on_played_ignores_non_episode_media_type(provider: VrtMaxProvider
 async def test_on_played_disabled_does_nothing(provider: VrtMaxProvider) -> None:
     """Without VRT credentials, progress is not written back."""
     provider._auth.enabled = False  # type: ignore[misc]
-    media_item = Mock(duration=1800)
+    media_item = _episode(duration=1800)
 
     await provider.on_played(
         MediaType.PODCAST_EPISODE,
@@ -213,7 +250,7 @@ async def test_on_played_swallows_errors(provider: VrtMaxProvider) -> None:
     provider._auth.get_access_token = AsyncMock(  # type: ignore[method-assign]
         side_effect=VrtAuthError("bad")
     )
-    media_item = Mock(duration=1800)
+    media_item = _episode(duration=1800)
 
     await provider.on_played(
         MediaType.PODCAST_EPISODE,
@@ -222,3 +259,59 @@ async def test_on_played_swallows_errors(provider: VrtMaxProvider) -> None:
         position=10,
         media_item=media_item,
     )
+
+
+class _FakeResponse:
+    """Minimal stand-in for an aiohttp response used as an async context manager."""
+
+    def __init__(self, payload: dict[str, Any]) -> None:
+        self.status = 200
+        self._payload = payload
+
+    async def json(self) -> dict[str, Any]:
+        return self._payload
+
+    async def __aenter__(self) -> Self:
+        return self
+
+    async def __aexit__(self, *_exc: object) -> None:
+        return None
+
+
+def _client_returning(payload: dict[str, Any]) -> VrtMaxClient:
+    """Return a client whose aggregator call yields the given payload."""
+    session = Mock()
+    session.get = Mock(return_value=_FakeResponse(payload))
+    return VrtMaxClient(session, logging.getLogger("test"))
+
+
+async def test_resolve_ondemand_hls_returns_the_drm_free_variant() -> None:
+    """The DRM-free rendition is picked out of the aggregator's target list."""
+    client = _client_returning(
+        {
+            "targetUrls": [
+                {"type": "hls", "url": "https://x/aud-1_drm_1.m3u8"},
+                {"type": "hls", "url": "https://x/aud-1_nodrm_1.m3u8"},
+            ]
+        }
+    )
+
+    assert await client.resolve_ondemand_hls("pub$aud", "ptok") == "https://x/aud-1_nodrm_1.m3u8"
+
+
+async def test_resolve_ondemand_hls_rejects_drm_only_response() -> None:
+    """With no DRM-free rendition on offer, resolution fails instead of returning one."""
+    client = _client_returning({"targetUrls": [{"type": "hls", "url": "https://x/drm.m3u8"}]})
+
+    # Returning the DRM target would need a decryption key we neither hold nor are
+    # entitled to, so playback would fail anyway, with a far less obvious reason.
+    with pytest.raises(MediaNotFoundError):
+        await client.resolve_ondemand_hls("pub$aud", "ptok")
+
+
+async def test_resolve_ondemand_hls_rejects_empty_target_list() -> None:
+    """An aggregator response with no HLS targets at all fails too."""
+    client = _client_returning({"targetUrls": []})
+
+    with pytest.raises(MediaNotFoundError):
+        await client.resolve_ondemand_hls("pub$aud", "ptok")
